Refactors code to follow SoC principle
