+++ basyx.components.databridge/src/test/java/basyx/components/databridge/regression/DataBridgeSuiteMqtt.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2022 the Eclipse BaSyx Authors
+ * Copyright (C) 2023 the Eclipse BaSyx Authors
  * 
  * Permission is hereby granted, free of charge, to any person obtaining
  * a copy of this software and associated documentation files (the
@@ -30,14 +30,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
-import javax.jms.Connection;
-import javax.jms.DeliveryMode;
-import javax.jms.Destination;
-import javax.jms.JMSException;
-import javax.jms.MessageProducer;
-import javax.jms.Session;
-import javax.jms.TextMessage;
-
 import org.apache.commons.io.FileUtils;
 import org.eclipse.basyx.aas.aggregator.api.IAASAggregator;
 import org.eclipse.basyx.aas.metamodel.api.IAssetAdministrationShell;
@@ -45,34 +37,33 @@ import org.eclipse.basyx.aas.metamodel.map.descriptor.CustomId;
 import org.eclipse.basyx.submodel.metamodel.api.ISubmodel;
 import org.eclipse.basyx.submodel.metamodel.api.identifier.IIdentifier;
 import org.eclipse.basyx.submodel.metamodel.api.submodelelement.ISubmodelElement;
+import org.eclipse.paho.client.mqttv3.IMqttClient;
+import org.eclipse.paho.client.mqttv3.MqttException;
+import org.eclipse.paho.client.mqttv3.MqttMessage;
+import org.eclipse.paho.client.mqttv3.MqttPersistenceException;
+import org.eclipse.paho.client.mqttv3.MqttSecurityException;
 import org.junit.Test;
 
 /**
  * Suite for testing that the Databridge is setup correctly
- * with ActiveMQ
+ * with Mqtt
  * 
  * @author danish
  *
  */
-public abstract class DatabridgeSuiteActiveMQ {
-	private static Connection connection;
-	private static Session session;
-	private static Destination destination;
-
+public abstract class DataBridgeSuiteMqtt {
 	protected static IIdentifier deviceAAS = new CustomId("TestUpdatedDeviceAAS");
-
-	protected abstract Connection getActiveMQConnection() throws JMSException;
-
+	
+	protected abstract IMqttClient getMqttClient() throws MqttException;
+	
 	protected abstract IAASAggregator getAASAggregatorProxy();
-
+	
 	@Test
-	public void datapointIsPropagated() throws Exception {
-		publishNewDatapoint("first-topic");
-
-		publishNewDatapoint("second-topic");
-
+	public void test() throws Exception {
+		publishNewDatapoint();
+		
 		waitForPropagation();
-
+		
 		checkIfPropertyIsUpdated();
 	}
 
@@ -82,57 +73,24 @@ public abstract class DatabridgeSuiteActiveMQ {
 
 	private void checkIfPropertyIsUpdated() throws InterruptedException {
 		IAASAggregator proxy = getAASAggregatorProxy();
-
+		
 		IAssetAdministrationShell aas = proxy.getAAS(deviceAAS);
 		ISubmodel sm = aas.getSubmodels().get("ConnectedSubmodel");
-
-		isPropertyValueUpdated(sm, "ConnectedPropertyA", "336.36");
-
-		isPropertyValueUpdated(sm, "ConnectedPropertyB", "858383");
-	}
-
-	private void isPropertyValueUpdated(ISubmodel sm, String id, String actualValue) {
-		ISubmodelElement updatedProp = sm.getSubmodelElement(id);
+		
+		ISubmodelElement updatedProp = sm.getSubmodelElement("ConnectedPropertyB");
 
 		Object propValue = updatedProp.getValue();
 		
-		assertEquals(actualValue, propValue);
+		assertEquals("858383", propValue);
 	}
 
-	private void publishNewDatapoint(String queueName) {
-		try {
-			configureAndStartConnection();
-			
-			session = connection.createSession(false, Session.AUTO_ACKNOWLEDGE);
-			
-			destination = session.createQueue(queueName);
-			
-			createProducerAndSendMessage();
-			
-			closeConnections();
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	private void createProducerAndSendMessage() throws JMSException, IOException {
-		MessageProducer producer = session.createProducer(destination);
-		producer.setDeliveryMode(DeliveryMode.NON_PERSISTENT);
-
+	private void publishNewDatapoint() throws MqttException, MqttSecurityException, MqttPersistenceException, IOException {
 		String json = FileUtils.readFileToString(new File(getClass().getClassLoader().getResource("test.json").getFile()), StandardCharsets.UTF_8);
 		
-		TextMessage message = session.createTextMessage(json);
-		
-		producer.send(message);
-	}
-
-	private void configureAndStartConnection() throws JMSException {
-		connection = getActiveMQConnection();
-		connection.start();
-	}
-	
-	private void closeConnections() throws JMSException {
-		session.close();
-		connection.close();
+		IMqttClient mqttClient = getMqttClient();
+		mqttClient.connect();
+		mqttClient.publish("PropertyB", new MqttMessage(json.getBytes()));
+		mqttClient.disconnect();
+		mqttClient.close();
 	}
 }
